Develop

+++ src/controllers/score/score.controller.ts
@@ -15,9 +15,9 @@ export class ScoreController {
   @Post()
   @UseInterceptors(
     FileFieldsInterceptor([{ name: 'file', maxCount: 1 }], upload))
-  importScore(@Body() request: any) {
+  importScore(@Request() req: any, @Body() request: any) {
     try {
-      return this.scoreService.importScore(request)
+      return this.scoreService.importScore(req.user.userId, request)
     } catch (err: any) {
       return {
         statusCode: err.statuscode,
@@ -41,9 +41,9 @@ export class ScoreController {
 
   @UseGuards(JwtAuthGuard)
   @Get('toAnnounce/:class_id')
-  getScoresToAnnounceByClass(@Param() param: getScoreDto) {
+  getScoresToAnnounceByClass(@Request() req: any, @Param() param: getScoreDto) {
     try {
-      return this.scoreService.getScoresToAnnounceByClass(param.class_id)
+      return this.scoreService.getScoresToAnnounceByClass(req.user.userId, param.class_id)
     } catch (err: any) {
       return {
         statusCode: err.statuscode,
@@ -55,9 +55,13 @@ export class ScoreController {
   @UseGuards(JwtAuthGuard)
   @Get('template/:class_id')
   @Header('Content-type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
-  async getScoreTemplate(@Param() param: getScoreDto, @Res() res: Response) {
+  async getScoreTemplate(@Request() req: any, @Param() param: getScoreDto, @Res() res: Response) {
     try {
-      const result = await this.scoreService.getScoreTemplate(param.class_id)
+      const result: any = await this.scoreService.getScoreTemplate(req.user.userId, param.class_id)
+      if (result.statusCode) {
+        res.send(result)
+        return res
+      }
       res.download(`${result}`, function (err) {
         fs.unlinkSync(`./${result}`)
       })
@@ -65,6 +69,7 @@ export class ScoreController {
         statusCode: 200,
         message: "success"
       }
+
     } catch (err: any) {
       return {
         statusCode: err.statuscode,
@@ -75,9 +80,9 @@ export class ScoreController {
 
   @UseGuards(JwtAuthGuard)
   @Patch()
-  async editScore(@Body() body: EditScoreDto) {
+  async editScore(@Request() req: any, @Body() body: EditScoreDto) {
     try {
-      return await this.scoreService.editScoreByScoreIdStdId(body)
+      return await this.scoreService.editScoreByScoreIdStdId(req.user.userId, body)
     } catch (err: any) {
       return {
         statusCode: err.statuscode,
@@ -88,9 +93,9 @@ export class ScoreController {
 
   @UseGuards(JwtAuthGuard)
   @Delete(':score_id')
-  async deleteScoreByScoreId(@Param() param: DeleteScoreDto) {
+  async deleteScoreByScoreId(@Request() req: any, @Param() param: DeleteScoreDto) {
     try {
-      return await this.scoreService.deleteScoreByScoreId(param.score_id)
+      return await this.scoreService.deleteScoreByScoreId(param.score_id, req.user.userId)
     } catch (err: any) {
       return {
         statusCode: err.statuscode,
